+++ src/Sulu/Bundle/ContactBundle/Resources/public/js/components/accounts/components/contacts/main.js
@@ -43,7 +43,7 @@ define([
 
             // back to list
             this.sandbox.on('sulu.header.back', function() {
-                this.sandbox.emit('sulu.contacts.accounts.list');
+                this.sandbox.emit('sulu.contacts.accounts.list', this.options.data);
             }, this);
 
             // add new record to datagrid